stock_picking_package_preparation_line: FIX headers, file and model names
