+++ stock_picking_package_preparation_line/models/__init__.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    About License, see __openerp__.py file
-#
-##############################################################################
+#    Author: Francesco Apruzzese
+#    Copyright 2015 Apulia Software srl
+#    Copyright 2015 Lorenzo Battistini - Agile Business Group
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 
 from . import ir_config
 from . import stock_picking_package_preparation_line
-from . import stock
+from . import stock_picking_package_preparation
+from . import stock_move
+from . import res_company
